Add HomeActivity
Adds HomeActivity and re-organizes folders
