+++ app/src/androidTest/java/cc/wordview/app/ui/screens/search/SearchTest.kt
@@ -26,6 +26,7 @@ import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performImeAction
 import androidx.compose.ui.test.performTextInput
 import cc.wordview.app.MainActivity
+import cc.wordview.app.ui.activities.home.composables.search.SearchRepository
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
 import org.junit.Before
