+++ app/src/main/java/cc/wordview/app/ui/activities/home/composables/Settings.kt
@@ -15,7 +15,7 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package cc.wordview.app.ui.screens
+package cc.wordview.app.ui.activities.home.composables
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
